Avoid passing registry down the entire tree as the first argument to evaluate
This refactors the usage of registry so that it's only passed in to the
NonTerminal grammar production where it's needed and avoids the cruft of
passing the redundant argument through every evaluate call.

This frees up the first argument of evaluate to be used for another
purpose-potentially injecting a hash of variable substitutions at
runtime.