Combine rules from template context map passed to generate
Add support for combining rules passed as a hash map to the generate
method. The simple use case for this is to provide dynamic expansion of
runtime values (ala ‘template variables’). There are other possibilities
for extending this towards being able to merge and combine grammars from
multiple sources or programmatically construct grammars dynamically.

This is an early draft of the feature. The API and merge semantics for rules may
change quite drastically once we’re played with this a bit more and some
of the flaws become more obvious.