Return a nested tree structure from recursive #evaluate calls
Instead of a flattened string, results are returned in a nested
structure which is a morphism to an s-expression so should be fairly
intuitive to understand.

This means that tree structures can be a valid output of the generator,
rather than just strings. It hints at future possibilities for generating symbolic story
structures or representations of graphical objects. So more than just
text, although the library is still focused on text for now.

It’s not unreasonable to surmise that this will also provide a small
performance improvement over the current implementation due to the
string concatenation being delayed until right at then end and thus fewer
intermediate objects being created in the midst of the `#generate` call.