Support memoized rules using the instance var sigil
By appending `@` to a rule reference, the result of the first
`#evaluate` call will be memoized and apply throught the grammar thereafter.

An incredibly useful feature for generating larger texts with specific
themes and references that need to be consistent throughout.

This is a very basic initial stab at this feature. Requires more
thorough testing and use in practice.