Fix non-deterministic tests
These tests were generating multiple output strings, based on assuming that
*eventually* all possible permutations of the generator would end up as
testable output. This assumption was incorrect and caused intermittent
failures in local development and on the CI server.

Instead, the tests are simple enough to use a single matcher with a
union of possible results. This is adequate for the context of testing
the YAML and JSON serialisation.