Provide hook for extending string modifiers using modules
Instead of providing custom mapping or filter functions within the
grammar, this hook allows separate modules to be embedded in a grammar
using a hook that mixes in the module in a similar way to how `include`
and `extend` works.

Test coverage needs to be improved and the naming conventions/API
probably needs to be tidied up and better defined, but the string
transform/modifier functionality is now roughly complete.