Support passing instances of Random down the evaluation tree
This fixes the issue with the Calyx internals messing with the global
random number generator for the whole Ruby process. Instead of calling
`Kernel.srand` for each grammar run, an instance of Random is created at
the top level and passed down the evaluation tree as a parameter.

Basic functionality should be verifiable by tests passing without
changing the input seeds.