Introduce new options format to grammar constructor
It’s now possible to pass configuration options as a hash to the grammar
on construction. This provides greater API consistency and allows
additional options to be added in future.

The legacy format of passing a random integer seed or random instance is
still supported. Not sure if this will be supported indefinitely or
deprecated and subsequently removed in a future release.