Fix mistake in the docs that mixed up monkeypatching and modifier APIs
This is most likely an artifact of copypasta. Monkeypatching was written
first and it’s likely that same code sample got reused to explain
modifiers—but the docs were never checked, so this has slipped through.

A useful followup here would be to test every block of code in the
README so that any additional problems like this can be fixed straight
away.