Internal refactoring to make rule definition traceable
In order to trace missing rule errors to specific lines in the input
grammar, rules need to be stored with a trace representing where the
rule was defined.

This provides the groundwork for tracing rule errors by extracting rule
definition into a separate class and using a factory/builder function to
generate rules, rather than store them as raw production rules. There
may be other ways to achieve this and the code could certainly do with
being tightened up a bit.