Introduce value object which encapsulates a generated result
As described in #23, this introduces a result object which wraps the
expression tree returned from evaluating a grammar and provides access
to the result as a raw tree, flattened string or symbol.

The generated result object is returned from a new
`Calyx::Grammar#generate_result` method. `Calyx::Grammar#generate` still
returns a string and `Calyx::Grammar#evaluate` now fires a deprecation
warning and will be removed soon.