Make unique rules cycle when the sequence is used up
Each unique substitution will reference a choice rule that has a
well-defined list of possible alternative results. Previously, the
choice picker would only work up until each result had been expanded
once. Referencing the rule again after this crashed the library with an
infinite hang.

With two possible options for the behaviour here—see issue #24—the
decision is either to throw an error or wrap around and cycle through
the values again. For clarity and simplicity, we’re going with the
second of those options until further notice.

See test spec. State is not kept between cycles so there’s no guarantee
that elements will come out in the same order each time. A possible
change to this feature would be to add a strict mode toggle that throws
an error rather than cycling. Unclear how useful this would be and I
want to get some feedback from how this gets used in practice before
locking it down in the API.