Deprecate YAML syntax for loading rules
YAML adds unncessary complexity via TMTOWTDI, which makes maintenance,
testing and porting to other languages more time consuming than it needs
to be. YAML was once a popular option for text encoding of data in Ruby
and Python but it’s waning in popularity and generally isn’t expected to
be used in languages like JavaScript that we want to target.

This changeset removes all non-breaking YAML details from the repo and
deprecates YAML loading with a warning message when loading rules. It’s
still possible to generate text from YAML but it’s not recommended and
the functionality will be removed in the near future.