Refactor rule AST construction
Moves AST for productions out of the registry so that changes to
serialization and DSL format can be centralised in the rule class.

This is a start on clearing the way for possible addition of attribute
grammar/indexed grammar features. Unclear how these will work at the AST
level as yet but one of the key interface changes is allowing
productions to be tagged via the DSL or JSON format. This may also mean
reconsidering the way arrays and hashes are used to define production
rules. Ideally these decisions can be explored and tested without having
to touch the main structure of the grammar and registry or the AST
evaluation chain.