Introduce bidirectional paired mapping rules
It’s now possible to declare grammar productions to be used as lookup
tables that can be applied as modifiers to rule expansions in template
expressions. A crude form of dynamic dispatch substitution, it allows
the result of rule expansions to be used as input to select other rules
rather than always getting concatenated into the result output.

It’s bidirectional to save repetition and add flexibility on larger text
projects. The dictionary lookup can be applied towards the right
(key => value) and towards the left (value => key).

The tests should explain largely what’s going on.

These changes are preliminary and don’t represent the final scope or
implementation of the feature.