Cleanup prefix tree and add wildcard string substitution
This is in no shape yet to release as a new feature but at least the
tests are all passing and a lot of the tricky internal code has a
default implementation. Whether this should be changed to a brute force
regex mapping in future is a separate question but there are some
advantages to doing it this way, such as being able to pop wildcard
search results directly out of the prefix tree.