Cleanup and refactor API after May prototyping
Moved onto some other projects very fast and left this repo in a bit of
a state earlier in the year. This refactor introduces several significant
changes into the overall architecture of Calyx.

- Original prototype ‘paired mapping’ is now ‘affix tree’
- Making a start on the design of namespaces and types separating
expression syntax and core grammar production types: ‘uniform branch’
replaces ‘choice’ and ‘weighted branch’ replaces ‘weighted choice’

The cleanup is mostly clearing stuff out of the spec files that
shouldn’t be there and getting the tests to pass properly.